docs: ✏️ Provide CONTRIBUTING.md file
Provide the necessary information for contributors in the
CONTRIBUTING.md file. Ensure the information provided is up to date and
easily understandable with clear format. This file should help everyone
willing to contribute to the project by describing on how to setup their
development environment and make the PR adhering to the project
guidelines.