docs: ✏️ Provide MAINTAINERS.md file
Provide the project management and issue tracking information for
project maintainers and contributors in the MAINTAINERS.md file. Ensure
that the information provided is up to date, matches the project labels
and is easily understandable with clear format. This file is intended to
help everyone understand the issue priority, status, release version as
well as the the overall project maintenance and management status.